removing the observables and simplifying the dropzone package
