+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@mainframework/file-upload",
-  "version": "1.0.0",
+  "version": "0.0.1",
   "description": "Package to handle the selection of files for upload",
   "main": "./dist/cjs/index.js",
   "module": "./dist/esm/index.js",
@@ -14,12 +14,14 @@
   },
   "homepage": "",
   "scripts": {
+    "build:css:dev": "npx tailwindcss -i ./src/index.css -o ./src/tailwind.css --watch",
+    "build:css:prod": "npx tailwindcss -i ./src/index.css -o ./src/tailwind.css --minify ",
     "husky": "husky",
     "clean": "rimraf dist",
     "lint": "eslint --ext 'ts,tsx' --max-warnings=0 --fix",
     "lint-staged": "lint-staged",
     "prebuild": "yarn husky && yarn clean",
-    "build": "yarn prebuild && yarn build:esm",
+    "build": "cross-env yarn prebuild && yarn build:css:prod && yarn build:esm",
     "build:esm": "tsc",
     "build:cjs": "tsc --module commonjs --outDir dist/cjs",
     "test": "jest --config jestconfig.json"
@@ -51,17 +53,20 @@
     ]
   },
   "devDependencies": {
+    "@tailwindcss/aspect-ratio": "^0.4.2",
+    "@tailwindcss/forms": "^0.5.7",
     "@testing-library/react": "^16.0.0",
     "@types/jest": "^29.5.12",
     "@types/react": "^18.3.3",
-    "@typescript-eslint/eslint-plugin": "^7.17.0",
-    "@typescript-eslint/parser": "^7.17.0",
+    "@typescript-eslint/eslint-plugin": "^7.18.0",
+    "@typescript-eslint/parser": "^7.18.0",
+    "cross-env": "^7.0.3",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.2.1",
     "eslint-plugin-react": "^7.35.0",
     "eslint-plugin-react-hooks": "^4.6.2",
-    "husky": "^9.1.3",
+    "husky": "^9.1.4",
     "jest": "^29.7.0",
     "jest-canvas-mock": "^2.5.2",
     "jest-environment-jsdom": "^29.7.0",
@@ -69,6 +74,7 @@
     "prettier": "^3.2.5",
     "react": "^18.3.1",
     "rimraf": "^6.0.1",
+    "tailwindcss": "^3.4.7",
     "ts-jest": "^29.2.3",
     "tslib": "^2.6.3",
     "typescript": "^5.5.4"
