+++ src/index.css
@@ -1,0 +1,29 @@
+@import "tailwindcss/base";
+@import "tailwindcss/components";
+@import "tailwindcss/utilities";
+
+/*
+    Any custom css, add it here
+*/
+
+@layer utilities {
+  /* Hide scrollbar for Chrome, Safari and Opera */
+  .no-scrollbar::-webkit-scrollbar {
+    display: none;
+  }
+  /* Hide scrollbar for IE, Edge and Firefox */
+  .no-scrollbar {
+    -ms-overflow-style: none; /* IE and Edge */
+    scrollbar-width: none; /* Firefox */
+  }
+}
+
+body {
+  @apply bg-indigo-500;
+}
+
+.glow-border {
+  border: 2px dotted transparent;
+  border-image: linear-gradient(45deg, cyan, magenta) 1;
+  box-shadow: 0 0 10px rgba(0, 255, 255, 0.7);
+}
