+++ src/index.ts
@@ -1,4 +1,1 @@
-export { useFileUploader } from "../src/shared/hooks/useFileUploader";
-export { FileUpload } from "../src/shared/components/FileUpload";
-export { type IFileUpload, type FileData, type ErrorMessage } from "../src/shared/types/types";
-export { type IObservable } from "../src/shared/utils/Observables";
+export { useFileSelector } from "./shared/hooks/useFileSelector";
