+++ src/shared/components/FileSelector/index.ts
@@ -1,0 +1,1 @@
+export { FileSelector } from "./FileSelector";
