--- src/shared/components/FileUpload/index.ts
@@ -1,1 +1,0 @@
-export { FileUpload } from "./FileUpload";
