+++ src/types/css.d.ts
@@ -1,0 +1,2 @@
+// Put this is for css modules to work
+declare module "*.css";
